+++ src/lib/config.ts
@@ -47,8 +47,8 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "4.17.0-alpha.1772253819150-d3f36156",
-  sdkVersion: "0.6.0-alpha.30",
-  genVersion: "2.845.15",
+  sdkVersion: "0.6.0-alpha.31",
+  genVersion: "2.846.1",
   userAgent:
-    "speakeasy-sdk/typescript 0.6.0-alpha.30 2.845.15 4.17.0-alpha.1772253819150-d3f36156 cribl-control-plane",
+    "speakeasy-sdk/typescript 0.6.0-alpha.31 2.846.1 4.17.0-alpha.1772253819150-d3f36156 cribl-control-plane",
 } as const;