+++ RELEASES.md
@@ -1229,3 +1229,13 @@ Based on:
 - [typescript v0.6.0-alpha.30] .
 ### Releases
 - [NPM v0.6.0-alpha.30] https://www.npmjs.com/package/cribl-control-plane/v/0.6.0-alpha.30 - .
+
+## 2026-03-04 02:11:41
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.739.0 (2.846.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.6.0-alpha.31] .
+### Releases
+- [NPM v0.6.0-alpha.31] https://www.npmjs.com/package/cribl-control-plane/v/0.6.0-alpha.31 - .